pyproject.toml
[tool.poetry]
name = "tele-bot"
version = "0.1.0"
description = ""
authors = ["EddiePhan <eddie.phan303@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.10.8"
aiogram = "^3.13.1"
python-telegram-bot = "13.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
